requirements.txt
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.2.0
einops
datasets
xformers
